Reduce memory usage by storing subprogram name information in two maps.
This is more efficient than the previous approach of trying to save a
reduced set of information for each subprogram DIE. The key observation
is that only one of the linkage name, the spec, or the name (in that
order) needs to be saved for each subprogram DIE, so it's simpler to
just maintain it in two maps. This also saves allocations when the
subprogram DIE contains no fields of interest.